geomeTRIC can request hessian calculation from QCEngine for TS optimizations (calc_cartesian_hessian)

+++ geometric/normal_modes.py
@@ -68,10 +68,11 @@ def calc_cartesian_hessian(coords, molecule, engine, dirname, read_data=True, ve
         Files are also written to <prefix.tmp>/hessian.
     """
     nc = len(coords)
+    dir = dirname is not None
     # Attempt to read existing Hessian data if it exists.
     # Read from files hessian.txt/coords.xyz, hessian_1.txt/coords_1.xyz, etc.
     counter = 0
-    while read_data:
+    while read_data and dir:
         if counter > 0:
             hesstxt = os.path.join(dirname, "hessian", "hessian_%i.txt" % counter)
             hessxyz = os.path.join(dirname, "hessian", "coords_%i.xyz" % counter)
@@ -91,20 +92,22 @@ def calc_cartesian_hessian(coords, molecule, engine, dirname, read_data=True, ve
         counter += 1
 
     # Compute hessian from scratch.
-    hesstxt = os.path.join(dirname, "hessian", "hessian.txt")
-    hessxyz = os.path.join(dirname, "hessian", "coords.xyz")
-    # First back up any existing Hessian data in a way that a future calculation could read it.
-    if os.path.exists(hessxyz) and os.path.exists(hesstxt):
-        bak(hessxyz)
-        bak(hesstxt)
+    if dir:
+        hesstxt = os.path.join(dirname, "hessian", "hessian.txt")
+        hessxyz = os.path.join(dirname, "hessian", "coords.xyz")
+        # First back up any existing Hessian data in a way that a future calculation could read it.
+        if os.path.exists(hessxyz) and os.path.exists(hesstxt):
+            bak(hessxyz)
+            bak(hesstxt)
 
-    oldxyz = molecule.xyzs[0].copy()
+    #oldxyz = molecule.xyzs[0].copy()
     molecule.xyzs[0] = coords.reshape(-1, 3)*bohr2ang
-    if not os.path.exists(os.path.join(dirname, "hessian")):
-        os.makedirs(os.path.join(dirname, "hessian"))
-        
-    molecule[0].write(hessxyz)
-    if not read_data: 
+    if dir:
+        if not os.path.exists(os.path.join(dirname, "hessian")):
+            os.makedirs(os.path.join(dirname, "hessian"))
+        molecule[0].write(hessxyz)
+
+    if not read_data and dir:
         if os.path.exists(os.path.join(dirname, "hessian", "displace")):
             shutil.rmtree(os.path.join(dirname, "hessian", "displace"))
     # Calculate Hessian using finite difference
@@ -134,6 +137,11 @@ def calc_cartesian_hessian(coords, molecule, engine, dirname, read_data=True, ve
             gbak = engine.read_wq(coords, dirname_d)['gradient']
             coords[i] += h
             Hx[i] = (gfwd-gbak)/(2*h)
+
+    elif type(engine).__name__ == "QCEngineAPI":
+        rec = engine.calc(coords, None)
+        Hx = rec["hessian"].reshape(nc, nc)
+
     else:
         # First calculate a gradient at the central point, for linking scratch files.
         engine.calc(coords, dirname, read_data=read_data)
@@ -152,12 +160,13 @@ def calc_cartesian_hessian(coords, molecule, engine, dirname, read_data=True, ve
     # Save Hessian to text file
     oldxyz = molecule.xyzs[0].copy()
     molecule.xyzs[0] = coords.reshape(-1, 3)*bohr2ang
-    molecule[0].write(hessxyz)
+    if dir:
+        molecule[0].write(hessxyz)
+        np.savetxt(hesstxt, Hx)
     molecule.xyzs[0] = oldxyz
-    np.savetxt(hesstxt, Hx)
     # Delete displacement calcs because they take up too much space
     keep_displace = False
-    if not keep_displace:
+    if not keep_displace and dir:
         if os.path.exists(os.path.join(dirname, "hessian", "displace")):
             shutil.rmtree(os.path.join(dirname, "hessian", "displace"))
     return Hx
